+++ src/main/java/cart/auth/AuthInterceptor.java
@@ -1,5 +1,6 @@
 package cart.auth;
 
+import cart.service.AuthService;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import org.springframework.beans.factory.annotation.Autowired;
